+++ src/base/internal_bnd.F90
@@ -62,7 +62,7 @@ contains
 
       class(cg_list),            intent(in) :: cglist !> the list on which to perform the boundary exchange
       integer,                   intent(in) :: ind    !> index of cg%q(:) 3d array
-      integer,         optional, intent(in) :: nb     !> number of grid cells to exchange (not implemented for comm3d)
+      integer(kind=4), optional, intent(in) :: nb     !> number of grid cells to exchange (not implemented for comm3d)
       integer(kind=4), optional, intent(in) :: dim    !> do the internal boundaries only in the specified dimension
 
       call internal_boundaries(cglist, ind, .true., nb, dim)
@@ -79,7 +79,7 @@ contains
 
       class(cg_list),            intent(in) :: cglist !> the list on which to perform the boundary exchange
       integer,                   intent(in) :: ind  !> index of cg%w(:) 4d array
-      integer,         optional, intent(in) :: nb   !> number of grid cells to exchange (not implemented for comm3d)
+      integer(kind=4), optional, intent(in) :: nb   !> number of grid cells to exchange (not implemented for comm3d)
       integer(kind=4), optional, intent(in) :: dim  !> do the internal boundaries only in the specified dimension
 
       call internal_boundaries(cglist, ind, .false., nb, dim)
@@ -112,7 +112,7 @@ contains
       class(cg_list),            intent(in) :: cglist !> the list on which to perform the boundary exchange
       integer,                   intent(in) :: ind    !> index of cg%q(:) 3d array or cg%w(:) 4d array
       logical,                   intent(in) :: tgt3d  !> .true. for cg%q, .false. for cg%w
-      integer,         optional, intent(in) :: nb     !> number of grid cells to exchange (not implemented for comm3d)
+      integer(kind=4), optional, intent(in) :: nb     !> number of grid cells to exchange (not implemented for comm3d)
       integer(kind=4), optional, intent(in) :: dim    !> do the internal boundaries only in the specified dimension
 
       integer                               :: g, d, n
